% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Probability of Passing General Upper and/or Lower Specification Limit}
\usage{
pp(Llim, Ulim, mu, sigma, n)
}
\arguments{
\item{Llim}{lower specification limit}

\item{Ulim}{upper specification limit}

\item{mu}{hypothetical mean of the attribute}

\item{sigma}{hypothetical standard deviation of the attribute}

\item{n}{sample size (number of locations)}
}
\value{
A numeric value of the passing/acceptance probability
}
\description{
The function for calculating the probability of passing a general upper and/or lower boundary.
}
\seealso{
\code{rl_pp} and \code{PPQ_pp}.
}
\author{
Yalin Zhu
}
