% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_pp.R
\name{rl_pp}
\alias{rl_pp}
\title{Probability of Passing Specification Test for a Release Batch}
\usage{
rl_pp(Llim, Ulim, mu, sigma, NV)
}
\arguments{
\item{Llim}{lower specification limit}

\item{Ulim}{upper specification limit}

\item{mu}{hypothetical mean of the attribute}

\item{sigma}{hypothetical standard deviation of the attribute}

\item{NV}{nominal volume for the specification test.}
}
\value{
A numeric value of the passing/acceptance probability
}
\description{
The function for calculating the probability of passing critical quality attributes (CQA) specification test .
}
\examples{
rl_pp(Llim=1.5, Ulim=3.5, mu=2.5, sigma=0.8)
}
\references{
Burdick, R. K., LeBlond, D. J., Pfahler, L. B., Quiroz, J., Sidor, L., Vukovinsky, K., & Zhang, L. (2017).
Statistical Applications for Chemistry, Manufacturing and Controls (CMC) in the Pharmaceutical Industry.
\emph{Springer}.
}
\seealso{
\code{PPQ_pp}, \code{pi_pp} and \code{ti_pp}.
}
\author{
Yalin Zhu
}
